filtering invoice done and delete product done also modified necessary functionality

+++ services/product.service.js
@@ -1,6 +1,6 @@
 const Product = require("../models/product.model")
 const Inventory = require("../models/inventory.model")
-exports.getProductService = async (filters, queries) => {
+exports.getProductService = async () => {
 
   const products = await Product.find({})
   return products;
@@ -29,8 +29,29 @@ exports.createProductService = async (data) => {
 }
 
 exports.deleteSingleProductService = async (id) => {
+  const findProduct = await Product.findById({ _id: id })
+  if (findProduct) {
+    const result = await Product.deleteOne({ _id: id });
+    if (!result.deletedCount) {
+      res.status(400).json({
+        success: false,
+        message: "Couldn't deleted Successfully."
+      })
+    }
+    const inventory = await Inventory.find({}).select('products').populate("products total_buy");
+    console.log(id)
+    const othersProduct = inventory[0].products.filter(prod => prod._id != id)
+    console.log(othersProduct)
 
-  const result = await Product.deleteOne({ _id: id });
-  return result;
+    // total Buy process
+    const totalBuy = inventory[0]?.total_buy - (findProduct.buy_price *
+      findProduct.current_quantity)
+    console.log(totalBuy)
+
+    await Inventory.updateOne({ _id: inventory[0]._id }, { products: othersProduct, total_buy: totalBuy })
+
+    return result;
+  }
+
 }
